renaming and course as example
