+++ src/Contexts/notifications.actions.jsx
@@ -1,0 +1,6 @@
+export const ERROR = 'ERROR'
+
+export const newError = (message) => ({
+	type: ERROR,
+	message
+})
